Remove unused .renderTarget{ComponentAlignment,PixelByteCost}
.renderTargetComponentAlignment = .colorRender.alignment
.renderTargetPixelByteCost = .colorRender.byteCost

The one exception to this is rg11b10ufloat, which is not renderable in
core, but has an extension for renderability. kTextureFormatInfo
currently only expresses core capabilities. The alignment and byteCost
for this format will need to be added back when the refactor is done to
make a table for "static" properties of texture formats (which never
vary based on capabilities) - shape of that refactor is TBD.